+++ app/Security/Dependency.php
@@ -21,6 +21,12 @@ class Dependency
 	 * Cache file name.
 	 */
 	const CACHE_FILE_NAME = ROOT_DIRECTORY . \DIRECTORY_SEPARATOR . 'cache' . \DIRECTORY_SEPARATOR . 'security' . \DIRECTORY_SEPARATOR . 'dependency.json';
+	/**
+	 * Symfony check url.
+	 *
+	 * @var string
+	 */
+	private $checkUrl = 'https://security.symfony.com/check_lock';
 
 	/**
 	 * Security checker.
@@ -35,7 +41,7 @@ class Dependency
 		if ($this->hasCache()) {
 			$result = $this->getCache();
 		} elseif (\App\RequestUtil::isNetConnection()) {
-			$result = (new \App\Utils\Dependencies())->check();
+			$result = $this->check();
 			$this->saveCache($result);
 		}
 		return $result;
@@ -75,5 +81,71 @@ class Dependency
 	private function getCache(): array
 	{
 		return \App\Json::decode(\file_get_contents(static::CACHE_FILE_NAME));
+	}
+
+	/**
+	 * Send lock file to verify.
+	 *
+	 * @return array
+	 */
+	public function check(): array
+	{
+		$result = [];
+		if (\App\RequestUtil::isNetConnection() && !empty($lockFile = $this->getLockFile(ROOT_DIRECTORY))) {
+			$options = \App\RequestHttp::getOptions();
+			$options['headers']['Content-Type'] = 'application/octet-stream';
+			$options['headers']['Content-Disposition'] = 'form-data; name="lock"; filename="composer.lock"';
+			$options['headers']['Accept'] = 'application/json';
+			$response = (new \GuzzleHttp\Client($options))->post($this->checkUrl, [
+				'body' => $lockFile
+			]);
+			$result = (array) \App\Json::decode($response->getBody());
+			$result = (\is_array($result) && !empty($result)) ? $result : [];
+		}
+		return $result;
+	}
+
+	/**
+	 * Get lock file content.
+	 *
+	 * @param string $lock
+	 *
+	 * @return string
+	 */
+	private function getLockFile($lock): string
+	{
+		if (is_dir($lock) && file_exists($lock . \DIRECTORY_SEPARATOR . 'composer.lock')) {
+			$lock = $lock . \DIRECTORY_SEPARATOR . 'composer.lock';
+		} elseif (preg_match('/composer\.json$/', $lock)) {
+			$lock = str_replace('composer.json', 'composer.lock', $lock);
+		}
+		return is_file($lock) ? $this->getLockContent($lock) : '';
+	}
+
+	/**
+	 * Get parsed lock file elements.
+	 *
+	 * @param string $lock
+	 *
+	 * @return string
+	 */
+	private function getLockContent($lock): string
+	{
+		$contents = json_decode(file_get_contents($lock), true);
+		$hash = $contents['content-hash'] ?? ($contents['hash'] ?? '');
+		$packages = ['content-hash' => $hash, 'packages' => []];
+		if (\is_array($contents['packages'])) {
+			foreach ($contents['packages'] as $package) {
+				$data = [
+					'name' => $package['name'],
+					'version' => $package['version'],
+				];
+				if (isset($package['time']) && false !== strpos($package['version'], 'dev')) {
+					$data['time'] = $package['time'];
+				}
+				$packages['packages'][] = $data;
+			}
+		}
+		return json_encode($packages);
 	}
 }
